# microservices/end2end-tests/src/main/openapi/iot-0.0.yml
openapi: 3.0.1
info:
  title: iot
  version: "0.0"
paths:
  /actuators:
    get:
      tags:
        - actuators
      operationId: list
      responses:
        "200":
          description: list 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Actuator"
    post:
      tags:
        - actuators
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActuatorCreateDto"
        required: true
      responses:
        "200":
          description: create 200 response
  /actuators/{id}:
    get:
      tags:
        - actuators
      operationId: get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: get 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Actuator"
    put:
      tags:
        - actuators
      operationId: update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActuatorCreateDto"
        required: true
      responses:
        "200":
          description: update 200 response
    delete:
      tags:
        - actuators
      operationId: delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: delete 200 response
  /actuators/{id}/state:
    patch:
      tags:
        - actuators
      operationId: updateState
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActuatorUpdateStateDto"
        required: true
      responses:
        "200":
          description: updateState 200 response
  /rooms:
    get:
      tags:
        - rooms
      operationId: list_1
      responses:
        "200":
          description: list_1 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      tags:
        - rooms
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomCreateDto"
        required: true
      responses:
        "200":
          description: create_1 200 response
  /rooms/{id}:
    get:
      tags:
        - rooms
      operationId: get_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: get_1 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    put:
      tags:
        - rooms
      operationId: update_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomCreateDto"
        required: true
      responses:
        "200":
          description: update_1 200 response
    delete:
      tags:
        - rooms
      operationId: delete_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: delete_1 200 response
  /rooms/{id}/actuators:
    get:
      tags:
        - rooms
      operationId: getActuators
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: getActuators 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Actuator"
  /rooms/{id}/sensors:
    get:
      tags:
        - rooms
      operationId: getSensors
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: getSensors 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sensor"
  /sensors:
    get:
      tags:
        - sensors
      operationId: list_2
      responses:
        "200":
          description: list_2 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sensor"
    post:
      tags:
        - sensors
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorCreateDto"
        required: true
      responses:
        "200":
          description: create_2 200 response
  /sensors/{id}:
    get:
      tags:
        - sensors
      operationId: get_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: get_2 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sensor"
    put:
      tags:
        - sensors
      operationId: update_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SensorCreateDto"
        required: true
      responses:
        "200":
          description: update_2 200 response
    delete:
      tags:
        - sensors
      operationId: delete_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: delete_2 200 response
components:
  schemas:
    Actuator:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        type:
          type: string
        targetState:
          type: string
    ActuatorCreateDto:
      type: object
      properties:
        roomId:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
        targetState:
          type: string
    ActuatorUpdateStateDto:
      type: object
      properties:
        targetState:
          type: string
    Room:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
    RoomCreateDto:
      type: object
      properties:
        name:
          type: string
    Sensor:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
        type:
          type: string
    SensorCreateDto:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        roomId:
          type: integer
          format: int64